[ALPS03595399] VPU: refine to support 32bit proj
[Detail] refine codes to support 32 bit project

[Solution] refine codes to support 32 bit projects

MTK-Commit-Id: a69fb38d89a4428e4bac98e3dc49cd658e513da7

CR-Id: ALPS03595399
Feature: MT6771 Platform Feature

+++ drivers/misc/mediatek/vpu/2.0/vpu_algo.c
@@ -264,7 +264,7 @@ int vpu_alloc_algo(struct vpu_algo **ralgo)
 		LOG_ERR("vpu_alloc_algo(), algo=0x%p\n", algo);
 		return -ENOMEM;
 	}
-	algo->info_ptr = (uint64_t) algo + sizeof(vlist_type(struct vpu_algo));
+	algo->info_ptr = (uintptr_t) algo + sizeof(vlist_type(struct vpu_algo));
 	algo->info_length = prop_info_data_length;
 
 	INIT_LIST_HEAD(vlist_link(algo, struct vpu_algo));
@@ -357,7 +357,7 @@ int vpu_dump_algo(struct seq_file *s)
 						  prop_desc->count,
 						  0);
 
-			info_data = (unsigned char *) (algo->info_ptr + prop_desc->offset);
+			info_data = (unsigned char *) ((uintptr_t)algo->info_ptr + prop_desc->offset);
 			memset(line_buffer, ' ', 24);
 			for (j = 0; j < data_length; j++, info_data++) {
 				int pos = j % 8;
